+++ src/main/java/club/kid7/bannermaker/util/InventoryUtil.java
@@ -3,7 +3,10 @@ package club.kid7.bannermaker.util;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
 
 public class InventoryUtil {
     /**
@@ -18,6 +21,37 @@ public class InventoryUtil {
         //若有放不進去的部分，直接噴在地上
         if (!itemsCanNotAddToInv.isEmpty()) {
             player.getWorld().dropItem(player.getLocation(), itemsCanNotAddToInv.get(0));
+        }
+    }
+
+    public static void sort(List<ItemStack> itemStacks) {
+        //移除空值
+        itemStacks.removeAll(Collections.singletonList(null));
+        //重新排序
+        itemStacks.sort((itemStack1, itemStack2) -> {
+            int c = Integer.compare(itemStack1.getType().ordinal(), itemStack2.getType().ordinal());
+            if (c == 0) {
+                c = -Integer.compare(itemStack1.getAmount(), itemStack2.getAmount());
+            }
+            return c;
+        });
+        //合併
+        ItemStack previous = null;
+        final Iterator<ItemStack> iterator = itemStacks.iterator();
+        while (iterator.hasNext()) {
+            final ItemStack item = iterator.next();
+            if (previous != null && previous.isSimilar(item) && previous.getAmount() < previous.getMaxStackSize()) {
+                int count = Math.min(item.getAmount(), previous.getMaxStackSize() - previous.getAmount());
+                if (count > 0) {
+                    previous.setAmount(previous.getAmount() + count);
+                    item.setAmount(item.getAmount() - count);
+                    if (item.getAmount() <= 0) {
+                        iterator.remove();
+                        continue;
+                    }
+                }
+            }
+            previous = item;
         }
     }
 }